riya-moun-data-analyst-agent/requirements.txt
fastapi
uvicorn
langchain
google-generativeai
pandas
numpy
seaborn
matplotlib
python-dotenv
